pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ah_look_at"
version = "1.0.0"
description = "Enables vision for examining images"
requires-python = ">=3.7"
dependencies = ["Pillow", "PyMuPDF", "opencv-python", "numpy"]


[project.optional-dependencies]
dev = ["pytest"]
